update deployment subgroup
